Move `SourceConfig` to `source_definition` module in dbt/artifacts
