+++ core/dbt/parser/sources.py
@@ -4,6 +4,7 @@ from typing import Iterable, Dict, Optional, Set, Any, List
 
 from dbt.adapters.capability import Capability
 from dbt.adapters.factory import get_adapter
+from dbt.artifacts.resources import SourceConfig
 from dbt.config import RuntimeConfig
 from dbt.context.context_config import (
     BaseContextConfigGenerator,
@@ -12,7 +13,6 @@ from dbt.context.context_config import (
 )
 from dbt.contracts.graph.components import FreshnessThreshold, Time
 from dbt.contracts.graph.manifest import Manifest, SourceKey
-from dbt.contracts.graph.model_config import SourceConfig
 from dbt.contracts.graph.nodes import (
     UnpatchedSourceDefinition,
     SourceDefinition,
